Check for SSD OLED app

+++ speedifyconf/net/usr/lib/speedifyconf/run.sh
@@ -15,12 +15,15 @@ run_speedify (){
    sleep 2
    ./speedify_cli startupconnect on > /dev/null
 
+  
    #OLED-SSD1306 status display (crude workaround for now)
-   sh /usr/lib/speedifyconf/rotatelog.sh &
-   sleep 1
-   cd /etc/ssd || exit 1
-   killall ssd 2> /dev/null
-   nice -n100 ssd -c config.json > /dev/null &
+   if [[ -f "/usr/bin/ssd" ]]; then
+    sh /usr/lib/speedifyconf/rotatelog.sh &
+    sleep 1
+    cd /etc/ssd || exit 1
+    killall ssd 2> /dev/null
+    nice -n100 ssd -c config.json > /dev/null &
+   fi
 }
 
 parse_apt_url(){
